+++ CHANGELOG.md
@@ -8,6 +8,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### BREAKING CHANGES
 
+- Changes to `@cumulus/aws-client/S3`
+  - The signature of the `getObjectSize` function has changed. It now takes a
+    params object with three properties:
+    - **s3**: an instance of an AWS.S3 object
+    - **bucket**
+    - **key**
+  - The `getObjectSize` function will no longer retry if the object does not exist
+
 - **CUMULUS-1969**
   - The `DiscoverPdrs` task now expects `provider_path` to be provided at
     `event.config.provider_path`, not `event.config.collection.provider_path`